Refactor so the hourly agent has its own Python class. Reference #37
